% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qassert.r
\name{reexported_qassert}
\alias{reexported_qassert}
\title{Re-export of qassert}
\usage{
reexported_qassert(x, rule, var.name = "x")
}
\arguments{
\item{x}{See \code{link[checkmate]{qassert}}.}

\item{rule}{See \code{link[checkmate]{qassert}}.}
}

