% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtest.r
\name{reexported_qtest}
\alias{reexported_qtest}
\title{Re-export of qtest}
\usage{
reexported_qtest(x, rule)
}
\arguments{
\item{x}{See \code{link[checkmate]{qtest}}.}

\item{rule}{See \code{link[checkmate]{qtest}}.}
}

